// ADCF/ClientApp/src/app/models/review.ts
export class Review {
    fullName: string;
    createdDate: string;
    startDate: string;
    endDate: string;
    reviewText: string;
    cleanlinessRating: number;
    comfortRating: number;
    locationRating: number;
    qualityRating: number;
    servicesRating: number;
    staffRating: number;

    createdTimeStamp: number;
    startTimeStamp: number;
    endTimeStamp: number;
}